+++ src/main/resources/app.props.build
@@ -1,5 +1,5 @@
 # =======================================================================
 # Ai-GenAI Service Properties
 # =======================================================================
-build.number=211
-build.date=Sat Jan 4 13:20:28 IST 2025
+build.number=212
+build.date=Sat Jan 4 14:31:35 IST 2025
